output answers to SSD1315 using second core

+++ lib/Day1/Day1.h
@@ -3,8 +3,8 @@
 
 #include <Arduino.h>
 
-int solve_day1_part1();
-int solve_day1_part2();
+long long solve_day1_part1();
+long long solve_day1_part2();
 
 void solve_day1();
