+++ lib/Day3/Day3.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 
-int solve_day3_part1();
+long long solve_day3_part1();
 long long solve_day3_part2();
 
 void solve_day3();
