+++ lib/Day5/Day5.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 
-int solve_day5_part1();
+long long solve_day5_part1();
 long long solve_day5_part2();
 
 void solve_day5();
