+++ lib/Day7/Day7.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 
-int solve_day7_part1();
+long long solve_day7_part1();
 long long solve_day7_part2();
 
 void solve_day7();
